Makes deleting history no longer delete starred urls. Thiseffectively reenables the code in ExpireHistoryBackend. I also madethe code consistent so that when we delete visits as the result ofhistory deletion we don't change the typed/visit count of theunderlying url.BUG=1214201 1256202TEST=none
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@1423 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/test/testing_profile.cc
@@ -5,6 +5,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "chrome/test/testing_profile.h"
 
+#include "chrome/browser/history/history_backend.h"
 #include "chrome/common/chrome_constants.h"
 
 TestingProfile::TestingProfile()
@@ -31,9 +32,8 @@ void TestingProfile::CreateHistoryService(bool delete_file) {
     file_util::AppendToPath(&path, chrome::kHistoryFilename);
     file_util::Delete(path, false);
   }
-  file_util::CreateDirectory(history_dir_);
-  history_service_ = new HistoryService();
-  history_service_->Init(GetPath());
+  history_service_ = new HistoryService(this);
+  history_service_->Init(GetPath(), bookmark_bar_model_.get());
 }
 
 void TestingProfile::DestroyHistoryService() {
@@ -57,11 +57,23 @@ void TestingProfile::DestroyHistoryService() {
   MessageLoop::current()->Run();
 }
 
-void TestingProfile::CreateBookmarkBarModel() {
-  std::wstring path = GetPath();
-  file_util::AppendToPath(&path, chrome::kBookmarksFileName);
-  file_util::Delete(path, false);
+void TestingProfile::CreateBookmarkBarModel(bool delete_file) {
+  // Nuke the model first, that way we're sure it's done writing to disk.
+  bookmark_bar_model_.reset(NULL);
+
+  if (delete_file) {
+    std::wstring path = GetPath();
+    file_util::AppendToPath(&path, chrome::kBookmarksFileName);
+    file_util::Delete(path, false);
+  }
   bookmark_bar_model_.reset(new BookmarkBarModel(this));
+  if (history_service_.get()) {
+    history_service_->history_backend_->bookmark_service_ =
+        bookmark_bar_model_.get();
+    history_service_->history_backend_->expirer_.bookmark_service_ =
+        bookmark_bar_model_.get();
+  }
+  bookmark_bar_model_->Load();
 }
 
 void TestingProfile::CreateTemplateURLModel() {
